test(e2e): realign golden path to reverse-trial activation (JOV-3757)
Jovie's pricing is a 14-day reverse trial with no card required at
signup, so the golden path's "first value" moment is the artist's
public profile going live with imported music, not a Stripe paywall.

Remove the Stripe checkout section from golden-path.spec.ts (pricing
poll, Pro=$39/3900 assertion, checkout-session poll, checkout.stripe.com
assertion) and replace STEP 8 with an anonymous-fan verification that
the public profile renders live: 200 response, artist name in the h1,
and the imported Spotify link visible.

Checkout coverage is relocated (not deleted) to billing-checkout.spec.ts,
which exercises the same Stripe checkout-session flow against the
standing E2E test user as a post-activation upgrade path.
